Refactored to add tests.
